HW-3 Create BaseContract to organize common functions - Create contract package
