+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/contract/BaseContract.kt
@@ -1,0 +1,9 @@
+package com.example.googryandroidarchitecturestudy.ui.contract
+
+interface BaseContract {
+    interface View {
+        fun hideKeyboard()
+    }
+
+    interface Presenter
+}
