+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/presenter/MoviePresenter.kt
@@ -2,7 +2,7 @@ package com.example.googryandroidarchitecturestudy.ui.presenter
 
 import com.example.googryandroidarchitecturestudy.data.repository.MovieRepository
 import com.example.googryandroidarchitecturestudy.domain.Movie
-import com.example.googryandroidarchitecturestudy.ui.MovieContract
+import com.example.googryandroidarchitecturestudy.ui.contract.MovieContract
 
 class MoviePresenter(
     private val view: MovieContract.View,
